pyproject.toml
[tool.poetry]
name = "open-o1"
version = "0.1.0"
description = ""
authors = ["tikendraw <tikendraksahu1029@gmail.com>"]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
litellm = "^1.46.4"
streamlit = "^1.38.0"
openai = "^1.46.0"
python-dotenv = "^1.0.1"
function-schema = "^0.4.2"
termcolor = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
